requirements.txt
streamlit
python-dotenv
agno
openai
